begin working on crazy amounts of vehicle models...

+++ AncientWarfare/src/shadowmage/ancient_warfare/common/vehicles/VehicleVarHelpers/CatapultVarHelper.java
@@ -94,7 +94,7 @@ public void readFromNBT(NBTTagCompound tag)
   this.crankAngle = tag.getFloat("cA");
   this.crankSpeed = tag.getFloat("cS");
   this.armAngle = tag.getFloat("aA");
-  this.armSpeed = tag.getLong("aS");
+  this.armSpeed = tag.getFloat("aS");
   }
 
 @Override
